consolidate oraclescriptrunner
